magit-reflog-mode: No longer derive from magit-log-mode
While a reflog can be shown using "git log -g --abbrev-commit
--pretty=online" a reflog is still rather different from a log.
Usually a reflog is shown using "git reflog show", which is an
alias for the mentioned log command and what Magit uses.

This also fixes a bug.  Previously when showing the real log for the
reflog entry at point the log arguments were picked up from the reflog
arguments used in the reflog buffer, which is undesirable because that
currently is just "-n256".