magit-reflog-limit: New option replacing magit-reflog-arguments
"-n" is the only reflog argument whose value the user can change.
Because of that it wouldn't be worth it to implement a refreshing
prefix.  The "+" key binding is already a more convenient way to
increase the limit when necessary.

Remove the `magit-reflog-arguments' option, which is a leftover
from a time when most such arguments (but not this one) were used
by `magit-popup' commands.